requirements.txt
streamlit>=1.39.0
python-dotenv>=1.0.1
pypdf>=4.2.0
google-genai>=0.5.0
google-adk>=0.2.0
aiohttp>=3.9.0
pydantic>=2.7.0
streamlit-webrtc>=0.47.0
audio-recorder-streamlit>=0.0.10
cachetools>=5.3.3
pytest>=8.3.0
